Adding packages json and readme

+++ mongoexportTSV.sh
@@ -1,8 +1,5 @@
 #!/usr/bin/env bash
-#Get the json file
-#echo "$@"	
-mongoexport --db rss --collection articles --out urls.json --type=json --fields link,summary,title
-
+parameters="$@"
 while [[ $# -gt 0 ]]
 do
 key="$1"
@@ -24,9 +21,16 @@ case $key in
 	fields=`echo $1 | sed -e 's/^[^=]*=//g'`
 	shift
 	;;
+    --type*)
+	echo "--type is not supported by the module, please run the script without --type"
+	exit 1
+	;;
 esac
 shift
 done
-echo "Json file exported, converting it to tsv"
-node convert.js $output $fields
-rm $output
+
+if `mongoexport $parameters --type=json` ; then
+	echo "Json file exported, converting it to tsv"
+	node convert.js $output $fields
+	rm $output
+fi
